The elements inside are located in Wrapper runner, and their associated stylings have the suffix -animate-button.
I'd like to fix the OptionsButton such that it matches the rest of the
stylings with the onhover=>onclick color changes. Right now it is only
onClick.

I added a new class to handle these changes, I'll try to finish that up
thursday.


The OptionsPanel is originally instantiated in TabPane, along with the
button that fires the even.